[desk-tool] Rename 'structure' to 'structure-builder' to avoid confusion
